+++ src/main.tsx
@@ -1,10 +1,21 @@
-import { StrictMode } from 'react'
+import { startTransition } from 'react'
 import { createRoot } from 'react-dom/client'
-import { App } from './App.tsx'
-import './index.css'
+import { createBrowserRouter, RouterProvider } from 'react-router'
+import { Provider } from 'react-redux'
 
-createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>
-)
+import { router } from './router'
+import { store } from './store'
+
+import './styles/main.scss'
+
+const root = document.getElementById('root')!
+
+const routes = createBrowserRouter(router)
+
+startTransition(() => {
+  createRoot(root).render(
+    <Provider store={store}>
+      <RouterProvider router={routes} />
+    </Provider>
+  )
+})