+++ src/pages/home.tsx
@@ -1,0 +1,7 @@
+import type { FC, JSX } from 'react'
+
+export const Component: FC = (): JSX.Element => {
+  return <h1>Home page</h1>
+}
+
+Component.displayName = 'HomePage'