+++ src/pages/index.tsx
@@ -1,0 +1,6 @@
+import type { FC, JSX } from 'react'
+import { Outlet } from 'react-router'
+
+export const Root: FC = (): JSX.Element => {
+  return <Outlet />
+}