+++ src/router.ts
@@ -1,0 +1,14 @@
+import { Root } from './pages'
+
+export const router = [
+  {
+    path: '/',
+    Component: Root,
+    children: [
+      {
+        path: '/',
+        lazy: () => import('./pages/home'),
+      },
+    ],
+  },
+]