+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.2.0 2025-08-06
+
+- add scss base config
+- add redux setup
+- add router setup
+
 ## v0.1.0 2025-08-05
 
 - initial release